feat: Redesign campaign creation to work with lists not individual prospects
UX Improvement - List-Based Campaign Creation:
- Show auto-selected list from Data Approval with checkmark
- Display other available lists as optional additions
- Filter out lists already assigned to other campaigns
- Show prospect counts per list
- Work at LIST level, not individual prospect level
- Total prospect count summary

This matches the user's requirement: 'the search that I am picking in the new campaigns screen needs to use that list for the campaign creation process. We dont want to see a list of prospects, it is unnecesary. We need to see the list that is already assigned to that campaign with a checkmark - and a list of other lists that are available (and if not alrady assigned to anther campaign)'